+++ frontend/src/components/feedback-form.tsx
@@ -1,19 +1,58 @@
+import React from "react";
 import ModalButton from "./buttons/ModalButton";
+import { request } from '#/services/api';
+import hotToast, { toast } from "react-hot-toast";
+
+const FEEDBACK_VERSION = "1.0";
+const VIEWER_PAGE = "https://www.all-hands.dev/share";
 
 interface FeedbackFormProps {
-  onSubmit: (permissions: "private" | "public", email: string) => void;
   onClose: () => void;
-  isSubmitting?: boolean;
+  polarity: "positive" | "negative";
 }
 
 export function FeedbackForm({
-  onSubmit,
   onClose,
-  isSubmitting,
+  polarity,
 }: FeedbackFormProps) {
-  const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
+  const [isSubmitting, setIsSubmitting] = React.useState(false);
+
+  const onPressToast = (password: string) => {
+    navigator.clipboard.writeText(password);
+    copiedToClipboardToast();
+  };
+
+  const shareFeedbackToast = (
+    message: string,
+    link: string,
+    password: string,
+  ) => {
+    hotToast(
+      <div className="flex flex-col gap-1">
+        <span>{message}</span>
+        <a
+          data-testid="toast-share-url"
+          className="text-blue-500 underline"
+          onClick={() => onPressToast(password)}
+          href={link}
+          target="_blank"
+          rel="noreferrer"
+        >
+          Go to shared feedback
+        </a>
+        <span onClick={() => onPressToast(password)} className="cursor-pointer">
+          Password: {password} <span className="text-gray-500">(copy)</span>
+        </span>
+      </div>,
+      { duration: 5000 },
+    );
+  };
+
+
+  const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
     event?.preventDefault();
     const formData = new FormData(event.currentTarget);
+    setIsSubmitting(true);
 
     const email = formData.get("email")?.toString();
     const permissions = formData.get("permissions")?.toString() as
@@ -21,8 +60,25 @@ export function FeedbackForm({
       | "public"
       | undefined;
 
-    console.log('submitting');
-    if (email) onSubmit(permissions || "private", email);
+    const feedback: Feedback = {
+      version: FEEDBACK_VERSION,
+      feedback: polarity,
+      email,
+      polarity,
+      permissions,
+    };
+
+    const response = await request('/api/submit-feedback', {
+      method: 'POST',
+      body: JSON.stringify(feedback),
+      headers: {
+        'Content-Type': 'application/json',
+      },
+    });
+    const { message, feedback_id, password } = response.body;
+    const link = `${VIEWER_PAGE}?share_id=${feedback_id}`;
+    shareFeedbackToast(message, link, password);
+    setIsSubmitting(false);
   };
 
   return (
