+++ openhands/server/data_models/feedback.py
@@ -10,7 +10,7 @@ from openhands.core.logger import openhands_logger as logger
 class FeedbackDataModel(BaseModel):
     version: str
     email: str
-    feedback: Literal['positive', 'negative']
+    polarity: Literal['positive', 'negative']
     permissions: Literal['public', 'private']
     trajectory: list[dict[str, Any]]
